requirements.txt
pandas
numpy
matplotlib
folium
streamlit
scikit-learn
pyyaml
python-dotenv